requirements.txt
python-telegram-bot>=21.8
python-dotenv>=1.0.0
